+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Adminer dev
+- Allow specifying operator in search anywhere
 
 ## Adminer 5.3.0 (released 2025-05-04)
 - Align numeric functions right